fix: categories enabled vs disabled
